+++ src/methods/$reverse/$reverse.d.ts
@@ -1,0 +1,5 @@
+import { $InputIterable, $GeneratorIterator } from '../../internal/$iterable';
+
+declare function $reverse<T = any>(iterable: $InputIterable<T>): $GeneratorIterator<T>;
+
+export default $reverse;